🚑 [Fix]: tsx로 수정 및 argTypes 수정(CustomArgs 추가)
